feat: main 페이지에서 데이터 요청 실패 시 토스트 띄우고 스플래시 제거
