+++ client/src/GalleryPage/ErrorPage.tsx
@@ -45,7 +45,7 @@ export default function ErrorPage() {
           <div className="modal error-modal">
             <span className="error-span">메인화면으로 돌아갑니다 ... {remainTime}</span>
             <button className="enter-sample-world-button" onClick={() => setUseSampleData(true)}>
-              샘플 월드로 입장하기
+              튜토리얼 입장하기
             </button>
           </div>
         </FullScreenModal>
